requirements.txt
appdirs==1.4.4
black==22.3.0
click==8.1.3
Mako==1.1.6
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.22.4
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.5.2
pydantic==1.9.1
pygubu==0.18.1
pygubu-designer==0.22
tomli==2.0.1
typing-extensions==4.2.0
